Dodanie dzielenia real

+++ c_examples/real_divide.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main(int argc, char **argv)
 {
-    float x = 12.3;
-    float b = 3.2;
-    float res = x / b;
+    double x = 12.3;
+    double b = 3.2;
+    double res = x / b;
 }
